requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.7
django-admin-filters==0.1.3
django-auditlog==3.0.0
django-grappelli==3.0.3
django-jazzmin==2.6.0
django-jet==1.0.8
django-simpleui==2025.1.13
django-suit==0.2.28
mysqlclient==2.2.7
pillow==11.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
